Fixes scroll crashes, move login spinner to left
Scrolling crashes: Float cannot be casted to Double

+++ JavaApp/src/main/java/org/lwjgl/glfw/GLFW.java
@@ -1089,7 +1089,7 @@ public class GLFW
                         break;
                     case CallbackBridge.EVENT_TYPE_SCROLL:
                         if (mGLFWScrollCallback != null) {
-                            mGLFWScrollCallback.invoke(ptr, (double)dataArr[1], (double)dataArr[2]);
+                            mGLFWScrollCallback.invoke(ptr, (float)dataArr[1], (float)dataArr[2]);
                         }
                         break;
                     case CallbackBridge.EVENT_TYPE_WINDOW_POS:
